Remove unnecessary shared content and add a file for sharing constants between App and UI tests
